Lazy load expensive explorer
Used rollup dependency visualizer to find the
biggest dependencies
