+++ frontend/.gitignore
@@ -11,6 +11,7 @@ node_modules
 dist
 dist-ssr
 *.local
+dependencies.html
 
 # Local config
 /public/config.js
