+++ frontend/src/components/global-explorer/full-explorer.tsx
@@ -1,192 +1,15 @@
-import '@xyflow/react/dist/base.css';
-
-import type { Node } from '@xyflow/react';
-import { Position, ReactFlowProvider, useReactFlow } from '@xyflow/react';
-import { Flex, theme } from 'antd';
-import { useCallback, useEffect, useState } from 'react';
-
-import { defaultFitViewOptions, NodeEditor } from '@/components/charts/node-editor/node-editor.tsx';
-import { CustomEdgeTypes, CustomNodeTypes } from '@/components/charts/node-editor/node-types.ts';
+import { ReactFlowProvider } from '@xyflow/react';
+import { lazy, Suspense } from 'react';
 import { LoadingSpinner } from '@/components/loading/loading-spinner/loading-spinner.tsx';
-import { useGetGraphDataQuery } from '@/store/features/graph/graph-api-slice.ts';
-import type { NodeContract } from '@/types/graph/graph-contract.ts';
-import { parseRegularNode } from '@/utils/node-parser.helper';
-import { LinkToDataOutputNode, LinkToDataProductNode, LinkToDatasetNode } from '../explorer/common';
-import styles from '../explorer/explorer.module.scss';
-import { parseEdges } from '../explorer/utils';
-import { parseDomainNode } from './nodes/domain-node-parser.helper';
-import { Sidebar, type SidebarFilters } from './sidebar/sidebar';
-import { useNodeEditor } from './use-node-editor';
 
-function parseFullNodes(nodes: NodeContract[], setNodeId: (id: string) => void, domainsEnabled = true): Node[] {
-    // Count how many children each domain node has
-    const childCounts = nodes
-        .filter((node) => node.type !== CustomNodeTypes.DomainNode)
-        .reduce((acc: Record<string, number>, node) => {
-            if (node.data.domain_id) {
-                acc[node.data.domain_id] = (acc[node.data.domain_id] || 0) + 1;
-            }
-            return acc;
-        }, {});
-
-    // Parse regular nodes
-    const regularNodes = nodes
-        .filter((node) => node.type !== CustomNodeTypes.DomainNode)
-        .map((node) => {
-            let extra_attributes = {};
-            switch (node.type) {
-                case CustomNodeTypes.DataProductNode:
-                    extra_attributes = {
-                        nodeToolbarActions: node.isMain ? null : <LinkToDataProductNode id={node.data.id} />,
-                        targetHandlePosition: Position.Left,
-                        assignments: node.data.assignments,
-                    };
-                    break;
-                case CustomNodeTypes.DatasetNode:
-                    extra_attributes = {
-                        nodeToolbarActions: node.isMain ? '' : <LinkToDatasetNode id={node.data.id} />,
-                        targetHandlePosition: Position.Right,
-                        targetHandleId: 'left_t',
-                    };
-                    break;
-                case CustomNodeTypes.DataOutputNode:
-                    extra_attributes = {
-                        nodeToolbarActions: node.isMain ? (
-                            ''
-                        ) : (
-                            <LinkToDataOutputNode id={node.id} product_id={node.data.link_to_id} />
-                        ),
-                        sourceHandlePosition: Position.Left,
-                        isActive: true,
-                        targetHandlePosition: Position.Right,
-                        targetHandleId: 'left_t',
-                    };
-                    break;
-                default:
-                    throw new Error(`Unknown node type: ${node.type}`);
-            }
-            return parseRegularNode(node, setNodeId, domainsEnabled, true, extra_attributes);
-        });
-
-    // Parse domain nodes (only if domains are enabled and they have children)
-    const domainNodes = domainsEnabled
-        ? nodes
-              .filter((node) => node.type === CustomNodeTypes.DomainNode && childCounts[node.id] > 0)
-              .map((node, index) => parseDomainNode(node, setNodeId, index))
-        : [];
-
-    // Domain nodes are parents so should come before their children in the array
-    const result = [...domainNodes, ...regularNodes];
-
-    return result;
-}
-
-function InternalFullExplorer() {
-    // Same as InternalExplorer but this one does not filter anything, it shows the full graph
-    // Also includes a sidebar to select nodes
-
-    const { edges, onEdgesChange, nodes, onNodesChange, onConnect, setNodes, setEdges, applyLayout } = useNodeEditor();
-    const currentInstance = useReactFlow();
-    const { token } = theme.useToken();
-    const [sidebarFilters, setSidebarFilters] = useState<SidebarFilters>({
-        dataProductsEnabled: true,
-        datasetsEnabled: true,
-        dataOutputsEnabled: true,
-        domainsEnabled: true,
-    });
-
-    useEffect(() => {
-        // Give React Flow time to update its internals
-        const timeout = setTimeout(() => {
-            currentInstance.fitView();
-            setNodeId(null); // Reset nodeId when the graph is updated
-        }, 50); // 50ms is usually enough
-
-        return () => clearTimeout(timeout);
-    }, [currentInstance]);
-
-    const [nodeId, setNodeId] = useState<string | null>(null);
-
-    const { data: graph, isFetching } = useGetGraphDataQuery(
-        {
-            includeDataProducts: sidebarFilters.dataProductsEnabled,
-            includeDatasets: sidebarFilters.datasetsEnabled,
-            includeDataOutputs: sidebarFilters.dataOutputsEnabled,
-            includeDomains: sidebarFilters.domainsEnabled,
-        },
-        {
-            skip: false,
-        },
-    );
-
-    const generateGraph = useCallback(async () => {
-        if (graph) {
-            const nodes = parseFullNodes(graph.nodes, setNodeId, sidebarFilters.domainsEnabled);
-            const edges = parseEdges(graph.edges, token);
-
-            // Explicitly specify straight edge so it doesn't default to default edge (which is a bezier curve)
-            const straightEdges = edges.map((edge) => ({
-                ...edge,
-                type: CustomEdgeTypes.StraightEdge,
-            }));
-
-            const positionedNodes = await applyLayout(nodes, straightEdges, sidebarFilters.domainsEnabled); // positions the nodes
-
-            setNodes(positionedNodes);
-            setEdges(straightEdges);
-        }
-    }, [graph, applyLayout, sidebarFilters, token, setEdges, setNodes]);
-
-    useEffect(() => {
-        generateGraph();
-    }, [generateGraph]);
-
-    if (isFetching) {
-        return <LoadingSpinner />;
-    }
-
-    // Custom node click handler
-    function handleNodeClick(_event: React.MouseEvent, node: Node) {
-        if (node)
-            setTimeout(() => {
-                currentInstance.fitView({
-                    ...defaultFitViewOptions,
-                    nodes: [node],
-                });
-            }, 50); // Small delay to let React Flow's internal logic complete, otherwise there seems to be an automatic (internally managed) fit view, but to the wrong position.
-    }
-
-    return (
-        <Flex className={styles.nodeWrapper}>
-            <Sidebar
-                nodes={nodes}
-                setNodes={setNodes}
-                onFilterChange={setSidebarFilters}
-                sidebarFilters={sidebarFilters}
-                nodeId={nodeId}
-                setNodeId={setNodeId}
-            />
-            <NodeEditor
-                nodes={nodes}
-                edges={edges}
-                onConnect={onConnect}
-                onNodesChange={onNodesChange}
-                onEdgesChange={onEdgesChange}
-                onNodeClick={handleNodeClick}
-                editorProps={{
-                    draggable: false,
-                    edgesReconnectable: false,
-                    nodesConnectable: false,
-                }}
-            />
-        </Flex>
-    );
-}
+const InternalFullExplorer = lazy(() => import('@/components/global-explorer/internal-explorer.tsx'));
 
 export function FullExplorer() {
     return (
         <ReactFlowProvider>
-            <InternalFullExplorer />
+            <Suspense fallback={<LoadingSpinner />}>
+                <InternalFullExplorer />
+            </Suspense>
         </ReactFlowProvider>
     );
 }
