+++ frontend/src/components/loading/loading-spinner/loading-spinner.module.scss
@@ -1,15 +1,6 @@
 @use '@/styles/globals';
 
 .loading-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
     height: 100%;
     width: 100%;
-
-    .loading {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
 }
