+++ frontend/src/pages/explorer/explorer.page.tsx
@@ -1,8 +1,6 @@
 import { Typography } from 'antd';
 import { useTranslation } from 'react-i18next';
-
-import { FullExplorer } from '@/components/global-explorer/full-explorer';
-
+import { FullExplorer } from '@/components/global-explorer/full-explorer.tsx';
 import styles from './explorer.page.module.scss';
 
 export function ExplorerPage() {
